api/requirements.txt
alembic==1.8.1
bcrypt==4.0.1
bidict==0.22.0
blinker==1.5
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-Seeder==1.2.0
flask-sock==0.5.2
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
protobuf==3.20.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.5
simple-websocket==0.8.1
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
wsproto==1.2.0
